added sharpe ratio

+++ src/main/java/optimizer/FinanceDataSource.java
@@ -10,7 +10,7 @@ public interface FinanceDataSource {
 
     public Boolean querySymbol(String symbol, Date startDate, Date endDate);
     public String getSymbol();
-    public List<Float> getAdjClose();
+    public List<Double> getAdjClose();
     public List<Integer> getVolume();
     public List<Date> getDates();
